+++ core-ng/src/main/java/core/framework/impl/db/Connections.java
@@ -10,12 +10,16 @@ import java.sql.SQLTimeoutException;
  * @author neo
  */
 final class Connections {
-    // refer to com.mysql.jdbc.integration.c3p0.MysqlConnectionTester, and standard sql state list:
-    // http://dev.mysql.com/doc/connector-j/en/connector-j-reference-error-sqlstates.html
-    // https://docs.oracle.com/cd/E16338_01/appdev.112/e10827/appd.htm#g642406
     static void checkConnectionState(PoolItem<Connection> connection, SQLException e) {
         String state = e.getSQLState();
-        if (state != null && state.startsWith("08")) {
+        // for state starts with "08"
+        // refer to com.mysql.jdbc.integration.c3p0.MysqlConnectionTester, and standard sql state list:
+        // http://dev.mysql.com/doc/connector-j/en/connector-j-reference-error-sqlstates.html
+        // https://docs.oracle.com/cd/E16338_01/appdev.112/e10827/appd.htm#g642406
+        // for state: S1009
+        // refer to com.mysql.cj.jdbc.exceptions.SQLExceptionsMapping.translateException() and com.mysql.cj.exceptions.MysqlErrorNumbers.SQL_STATE_ILLEGAL_ARGUMENT
+        // MySQL jdbc connector will translate "statement is closed" and other errors to "S1009"
+        if (state != null && (state.startsWith("08") || "S1009".equals(state))) {
             connection.broken = true;
         }